AIDA-747 - Change importance from integer to double. Trying to figure out why double value of 9.999999e6 outputs in the model as:    aida:importance  "9999999.0"^^xsd:double ; but double value of 1.000001e7 outputs in the model (without the type) as:    aida:importance  1.000001E7 ;
See ExamplesAndValidationTest.simpleHypothesisWithImportanceWithCluster()

+++ python/aida_interchange/aifutils.py
@@ -401,7 +401,7 @@ def mark_importance(g, resource, importance):
     """
     Mark resource as having an importance value
     """
-    g.add((resource, AIDA_ANNOTATION.importance, Literal(importance, datatype=XSD.int)))
+    g.add((resource, AIDA_ANNOTATION.importance, Literal(importance, datatype=XSD.double)))
 
 
 def mark_informative_justification(g, resource, informative_justification):
