+++ src/main/java/com/ncc/aif/AIFUtils.java
@@ -955,7 +955,7 @@ public class AIFUtils {
      * @param resource   The Resource to mark with the specified importance
      * @param importance The importance value with which to mark the specified Resource
      */
-    public static void markImportance(Resource resource, Integer importance) {
+    public static void markImportance(Resource resource, Double importance) {
         resource.addLiteral(AidaAnnotationOntology.IMPORTANCE_PROPERTY, importance);
     }
